feat: 마이페이지 프로필 조회

+++ server/src/main/java/project/developer_pacemaker/repository/UserRepository.java
@@ -5,8 +5,10 @@ import org.springframework.stereotype.Repository;
 import project.developer_pacemaker.entity.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByEmail(String email);
+
+//    UserEntity findByUSeq(Long uSeq);
 
     Boolean existsByEmail(String email);
